test(Gallery): separate tests by files, rename setup function and remove timers mocks to fake timers api
