+++ packages/vkui/src/components/Gallery/Gallery.tsx
@@ -38,7 +38,6 @@ export const Gallery = ({
   const handleChange: GalleryProps['onChange'] = React.useCallback(
     (current: number) => {
       if (current === slideIndex) {
-        /* istanbul ignore next: не придумал как воспроизвести */
         return;
       }
       !isControlled && setSlideIndex(current);
